Fix connections to ProPresenter on Windows
Yes these fixes are only necessary for connections to ProPresenter running on windows.

- Action `presentationTriggerIndex`:
For every quick bible slide clicked, this Action will be invoked.
The `data.presentationPath` is incorrectly the presentation-path of the last "real" presentation.

- Action `clearText`:
For every quick bible slide clicked, this Action will be invoked. This is wrong.

- Action `fv`:
UIDs of current and next slide are not reliable. They change every time. WTF.